+++ src/js/components/Shortcut/index.tsx
@@ -3,8 +3,8 @@ import HotkeyDialog from './HotkeyDialog'
 import Hint from './Hint'
 
 export default () => (
-  <React.Fragment>
+  <>
     <HotkeyDialog key='dialog' />
     <Hint key='hint' />
-  </React.Fragment>
+  </>
 )